refactor: Height 클래스 불변성 유지하도록 변경
