+++ src/main/java/controller/LadderGameController.java
@@ -3,7 +3,7 @@ package controller;
 import domain.LadderGame;
 import domain.ladder.Height;
 import domain.ladder.Ladder;
-import domain.participants.Participants;
+import domain.participant.Participants;
 import domain.result.Prizes;
 import java.util.List;
 import utils.RandomStepGenerator;
