+++ src/main/java/domain/participant/Participants.java
@@ -1,4 +1,4 @@
-package domain.participants;
+package domain.participant;
 
 import java.util.List;
 import java.util.stream.IntStream;
