+++ src/main/java/view/OutputView.java
@@ -3,8 +3,8 @@ package view;
 import domain.ladder.Ladder;
 import domain.ladder.Line;
 import domain.ladder.StepPoint;
-import domain.participants.Name;
-import domain.participants.Participants;
+import domain.participant.Name;
+import domain.participant.Participants;
 import domain.result.Prize;
 import domain.result.Prizes;
 import java.util.List;
