+++ src/test/java/domain/PrizeTest.java
@@ -3,7 +3,7 @@ package domain;
 import static domain.result.Prizes.MAX_OF_PRIZE_LENGTH;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.participants.Participants;
+import domain.participant.Participants;
 import domain.result.Prizes;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
