+++ src/test/java/domain/PrizesTest.java
@@ -3,7 +3,7 @@ package domain;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import domain.participants.Participants;
+import domain.participant.Participants;
 import domain.result.Prizes;
 import java.util.List;
 import org.junit.jupiter.api.DisplayName;
